pyproject.toml
[tool.poetry]
name = "carnoises2"
version = "0.1.0"
description = ""
authors = ["Anatoly Levchik <endight@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
google-api-python-client = "^1.10.0"
pandas = "^1.1.1"
tqdm = "^4.48.2"
baker = "^1.3"
skorch = "^0.8.0"
resampy = "^0.2.2"
torch = "^1.6.0"
soundfile = "^0.10.3"
pytorch-ignite = "^0.4.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
